+++ src/api/data_types/chunking/hash_algorithm.rs
@@ -1,0 +1,7 @@
+use serde::Deserialize;
+
+#[derive(Debug, Deserialize, Clone, Copy, Eq, PartialEq, Ord, PartialOrd)]
+pub enum ChunkHashAlgorithm {
+    #[serde(rename = "sha1")]
+    Sha1,
+}
