+++ src/api/data_types/chunking/mod.rs
@@ -1,0 +1,16 @@
+//! Data types used in the API for sending and receiving data
+//! from the server.
+
+mod artifact;
+mod compression;
+mod dif;
+mod file_state;
+mod hash_algorithm;
+mod upload;
+
+pub use self::artifact::{AssembleArtifactsResponse, ChunkedArtifactRequest};
+pub use self::compression::ChunkCompression;
+pub use self::dif::{AssembleDifsRequest, AssembleDifsResponse, ChunkedDifRequest};
+pub use self::file_state::ChunkedFileState;
+pub use self::hash_algorithm::ChunkHashAlgorithm;
+pub use self::upload::{ChunkUploadCapability, ChunkUploadOptions};
